+++ CHANGELOG.md
@@ -28,7 +28,7 @@ Commands are explained like so:
     ```
 
 ### Fixed
-- none
+- Fixed `/modping` command to use moderators roles from the `config.json` file instead of using the role name, removed duplicated code.
 
 ## [v2.4.0] 15/05/2022